EDSC-3423: Formats subscription query tooltip
