+++ static/src/js/components/Subscriptions/SubscriptionsListItem.scss
@@ -42,34 +42,10 @@
     width: 20rem;
   }
 
-  &__tooltip-query-list {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-  }
-
   &__tooltip-query-heading {
     text-align: left;
     margin-bottom: 0.325rem;
     padding-bottom: 0.325rem;
     border-bottom: 1px solid rgba($color__white, 0.15);
-  }
-
-  &__tooltip-query-list-item {
-    display: flex;
-    margin-bottom: 0.125rem;
-  }
-
-  &__tooltip-query-list-item-heading {
-    flex-shrink: 0;
-    flex-grow: 0;
-    margin-right: 0.125rem;
-    font-size: 0.75rem;
-  }
-
-  &__tooltip-query-list-item-value {
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
   }
 }
