+++ static/src/js/util/humanizedQueryValueFormattingMap.js
@@ -27,20 +27,20 @@ export const platformsHierarchy = [
  * the values to be displayed in UI
  */
 export const humanizedQueryValueFormattingMap = {
-  boundingBox: (value) => formatPoints(value),
+  boundingBox: (value) => formatPoints([].concat(value)),
   browseOnly: (value) => formatBoolean(value),
-  circle: (value) => formatCircle(value),
+  circle: (value) => formatCircle([].concat(value)),
   cloudHosted: (value) => formatBoolean(value),
   dataCenterH: (value) => formatFacetHierarchy(value),
   granuleDataFormatH: (value) => formatFacetHierarchy(value),
   horizontalDataResolutionRange: (value) => formatFacetHierarchy(value),
   instrumentH: (value) => formatFacetHierarchy(value),
   latency: (value) => formatFacetHierarchy(value),
-  line: (value) => formatPoints(value),
+  line: (value) => formatPoints([].concat(value)),
   onlineOnly: (value) => formatBoolean(value),
   platformsH: (value) => formatFacetHierarchy(value, platformsHierarchy),
-  point: (value) => formatPoints(value),
-  polygon: (value) => formatPoints(value),
+  point: (value) => formatPoints([].concat(value)),
+  polygon: (value) => formatPoints([].concat(value)),
   processingLevelIdH: (value) => formatFacetHierarchy(value),
   projectH: (value) => formatFacetHierarchy(value),
   scienceKeywordsH: (value) => formatFacetHierarchy(value, scienceKeywordHierarchy),
